# src/main/java/com/pratikdesai/transfers/dto/response/ErrorResponse.java
package com.pratikdesai.transfers.dto.response;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.Instant;
import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@JsonInclude(JsonInclude.Include.NON_NULL)
public class ErrorResponse {

    @JsonProperty("error")
    private String error;

    @JsonProperty("message")
    private String message;

    @JsonProperty("details")
    private List<String> details;

    @JsonProperty("timestamp")
    private Instant timestamp;

    public static ErrorResponse of(String error, String message) {
        return ErrorResponse.builder()
                .error(error)
                .message(message)
                .timestamp(Instant.now())
                .build();
    }

    public static ErrorResponse of(String error, String message, List<String> details) {
        return ErrorResponse.builder()
                .error(error)
                .message(message)
                .details(details)
                .timestamp(Instant.now())
                .build();
    }
}